♻️ refact(core): migration complète des services Service/* vers architecture métier
- Déplacement et renommage de tous les services Service/* dans les dossiers métier (Uploader, Photo, File, Archive)
- Correction de tous les imports, namespaces et signatures
- Suppression des anciens fichiers Service/* et tests orphelins
- Restauration de FileNameGeneratorInterface dans Uploader/
- Tous les tests unitaires et d’intégration passent hors cas métier connu

BREAKING CHANGE: plus aucun service dans src/Service, tous les usages doivent pointer vers les nouveaux namespaces.
