--- src/Service/UploadDirectoryManager.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Service;
-
-class UploadDirectoryManager
-{
-    public function ensureDirectoryExists(string $directory): void
-    {
-        if (!is_dir($directory)) {
-            if (!@mkdir($directory, 0775, true) && !is_dir($directory)) {
-                throw new \RuntimeException("Impossible de créer le répertoire d'upload : $directory");
-            }
-        }
-        if (!is_writable($directory)) {
-            throw new \RuntimeException("Le répertoire d'upload n'est pas accessible en écriture : $directory");
-        }
-    }
-}
