+++ tests/Unit/ExifExtractorTest.php
@@ -2,7 +2,7 @@
 
 namespace App\Tests\Unit;
 
-use App\Service\ExifExtractor;
+use App\Photo\ExifExtractor;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
